requirements.txt
fastapi
uvicorn
motor
python-dotenv
pyjwt
passlib[bcrypt]
pydantic
python-multipart
dnspython